third_party/blink/web_tests/fast/forms/file/file-input-webkitdirectory-click-expected.txt

CONSOLE MESSAGE: line 5: FileChooser: opened; mode=kUploadFolder
Ensure clicking on an INPUT element with TYPE=FILE and WEBKITDIRECTORY launches a file chooser. Automated test passes if 'FileChooser: opened' was logged.
